# openapi.yaml
openapi: 3.0.1
info:
  title: Actual relay API
  version: 0.1.0
  description: ActualBudget (https://actualbudget.com) relay API. It describes rienafairefr/actual-relay, which connects to a locally running Actual app (version 0.0.144)
paths:
  /query:
    post:
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
      requestBody:
        content:
          text/plain:
            schema:
              type: string
  /budget-months:
    get:
      responses:
        default:
          description: "Budget-months response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMonths"
  /budget-load:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                budgetId:
                  type: string
      responses:
        204:
          description: OK
        400:
          description: NOK
  /budget/{month}:
    parameters:
      - name: month
        in: path
        required: true
        schema:
          type: string
    get:
      responses:
        default:
          description: "Budget-months response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMonthResponse"
  /budget/{month}/{categoryId}:
    parameters:
      - name: month
        in: path
        required: true
        schema:
          type: string
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
    patch:
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - type: object
                  properties:
                    amount:
                      type: number
                - type: object
                  properties:
                    carryover:
                      type: number
      responses:
        default:
          description: "Budget-month category patch response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMonthCategoryResponse"
  /accounts/{accountId}/transactions-import:
    parameters:
      - name: accountId
        required: true
        in: path
        schema:
          type: string
    post:
      summary: Import transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        default:
          description: " response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMonths"
  /accounts/{accountId}/transactions:
    parameters:
      - in: path
        name: accountId
        required: true
        schema:
          type: string
          format: uuid
    get:
      parameters:
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          required: false
      responses:
        default:
          description: "Account transactions response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountTransactions"
  /accounts/{accountId}/filtered-transactions:
    parameters:
      - in: path
        name: accountId
        required: true
        schema:
          type: string
          format: uuid
    get:
      parameters:
        - in: query
          schema:
            type: string
          name: text
          required: true
      responses:
        default:
          description: "Account transactions response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountTransactions"
    post:
      summary: Add a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        default:
          description: "Budget-months response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetMonths"
  /transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      responses:
        204:
          description: OK
        400:
          description: NOK
      requestBody:
        content:
          application/json:
            schema:
              type: object
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
  /accounts:
    get:
      responses:
        default:
          description: "accounts response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                type: string
                format: uuid
  /accounts/{id}":
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
    patch:
      responses:
        204:
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
  /accounts/{id}/close:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    post:
      responses:
        204:
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
  /accounts/{id}/reopen:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    post:
      responses:
        204:
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
  /category-groups:
    get:
      responses:
        default:
          description: "Category Groups"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryGroups"

    post:
      responses:
        204:
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
  /category-groups/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        204:
          description: OK
        400:
          description: NOK
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
  /categories:
    get:
      responses:
        default:
          description: "Categories response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesResponse"
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        204:
          description: OK
        400:
          description: NOK
  /categories/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        204:
          description: OK
        400:
          description: NOK
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
  /payees:
    get:
      responses:
        default:
          description: "Payees"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayeesResponse"
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        201:
          description: OK
        400:
          description: NOK
  /payees/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        204:
          description: OK
        400:
          description: NOK
    get:
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payee"
        400:
          description: NOK
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
  /payees/{payeeId}/rules:
    parameters:
      - in: path
        name: payeeId
        required: true
        schema:
          type: string
    get:
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayeeRule"
        400:
          description: NOK
    patch:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        204:
          description: OK
        400:
          description: NOK
    delete:
      responses:
        204:
          description: OK
        400:
          description: NOK
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayeeRule"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayeeRule"
        400:
          description: NOK

components:
  schemas:
    AccountTransactions:
      type: array
      items:
        $ref: "#/components/schemas/Transaction"
    CategoriesResponse:
      type: array
      items:
        $ref: "#/components/schemas/Category"
    BudgetMonthCategoryResponse:
      type: array
      items:
        type: object
    BudgetMonths:
      type: array
      items:
        type: string
    Category:
      type: object
      required:
        - name
        - group_id
      properties:
        group_id:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        is_income:
          type: boolean
        name:
          type: string
    Account:
      type: object
      properties:
        closed:
          type: boolean
        id:
          type: string
          format: uuid
        name:
          type: string
        offbudget:
          type: boolean
        type:
          type: string
          enum:
            [
              "checking",
              "savings",
              "investment",
              "credit",
              "mortgage",
              "debt",
              "other",
            ]
    CreateAccountRequest:
      type: object
      required:
        - account
      properties:
        account:
          $ref: "#/components/schemas/CreateAccount"
        initialBalance:
          type: number
    CreateAccount:
      type: object
      properties:
        closed:
          type: boolean
        name:
          type: string
        offbudget:
          type: boolean
        type:
          type: string
          enum:
            [
              "checking",
              "savings",
              "investment",
              "credit",
              "mortgage",
              "debt",
              "other",
            ]
    Accounts:
      type: array
      items:
        $ref: "#/components/schemas/Account"
    QueryResponse:
      type: object
    PayeesResponse:
      type: array
      items:
        $ref: "#/components/schemas/Payee"
    BudgetCategory:
      type: object
      properties:
        balance:
          type: number
        budgeted:
          type: number
        carryover:
          type: boolean
        group_id:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        is_income:
          type: boolean
        name:
          type: string
        spent:
          type: number
    CategoryGroups:
      type: array
      items:
        $ref: "#/components/schemas/CategoryGroup"
    PayeeRule:
      type: object
      required:
        - type
        - value
      properties:
        id:
          type: string
          format: uuid
        payee_id:
          type: string
          format: uuid
        type:
          type: string
          enum: ["equals", "contains"]
          description: Must be one of equals or contains
        value:
          type: string
          description: Value to match imported payee names on
    CategoryGroup:
      type: object
      required:
        - name
      properties:
        balance:
          type: number
        budgeted:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategory"
        id:
          type: string
          format: uuid
        is_income:
          type: boolean
        name:
          type: string
        spent:
          type: number
    BudgetMonthResponse:
      type: object
      properties:
        categoryGroups:
          type: array
          items:
            $ref: "#/components/schemas/CategoryGroup"
        forNextMonth:
          type: number
        fromLastMonth:
          type: number
        incomeAvailable:
          type: number
        lastMonthOverspent:
          type: number
        month:
          type: string
        toBudget:
          type: number
        totalBalance:
          type: number
        totalBudgeted:
          type: number
        totalIncome:
          type: number
        totalSpent:
          type: number
    CreateTransactionRequest:
      type: array
      items:
        $ref: "#/components/schemas/CreateTransaction"
    Transaction:
      type: object
      required:
        - account
        - date
      properties:
        id:
          type: string
          format: uuid
        account:
          type: string
          format: uuid
        error:
          type: string
          nullable: true
        date:
          type: string
          format: date
        amount:
          type: number
        payee:
          type: string
          format: uuid
          description: If given, this overrides payee_name.
          nullable: true
        payee_name:
          type: string
          description: If given, a payee will be created with this name. If this matches an already existing payee, it will use it.
        imported_payee:
          type: string
          description: This can be anything. Meant to represent the raw description when importing, allowing the user to see the original value.
          nullable: true
        imported_id:
          type: string
          description: A unique id usually given by the bank, if importing. Use this is avoid duplicate transactions.
          nullable: true
        parent_id:
          type: string
          nullable: true
        is_child:
          type: boolean
        sort_order:
          type: number
        starting_balance_flag:
          type: boolean
        tombstone:
          type: boolean
        is_parent:
          type: boolean
        category:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          nullable: true
        transfer_id:
          type: string
          description: If a transfer, the id of the transaction in the other account for the transfer. See transfers.
          nullable: true
        cleared:
          type: boolean
          description: A flag indicating if the transaction has cleared or not.
        schedule:
          type: string
          nullable: true
        subtransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
    CreateTransaction:
      type: object
      required:
        - account
        - date
      properties:
        id:
          type: string
          format: uuid
        account:
          type: string
          format: uuid
        date:
          type: string
          format: date
        amount:
          type: number
        payee:
          type: string
          format: uuid
          description: If given, this overrides payee_name.
          nullable: true
        payee_name:
          type: string
          description: If given, a payee will be created with this name. If this matches an already existing payee, it will use it.
        imported_payee:
          type: string
          description: This can be anything. Meant to represent the raw description when importing, allowing the user to see the original value.
          nullable: true
        imported_id:
          type: string
          description: A unique id usually given by the bank, if importing. Use this is avoid duplicate transactions.
          nullable: true
        category:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          nullable: true
        transfer_id:
          type: string
          description: If a transfer, the id of the transaction in the other account for the transfer. See transfers.
          nullable: true
        cleared:
          type: boolean
          description: A flag indicating if the transaction has cleared or not.
        schedule:
          type: string
          nullable: true
        subtransactions:
          type: array
          items:
            $ref: "#/components/schemas/CreateTransaction"
    Payee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          nullable: true
        transfer_acct:
          type: string
          nullable: true